refactor: 根据代码规范重构
