+++ Scripts/2D/Character/CharacterCreator.cs
@@ -48,17 +48,4 @@
             return g;
         }
     }
-
-    /// <summary>
-    /// 角色创建器基
-    /// </summary>
-    public interface ICharacterCreator
-    {
-        /// <summary>
-        /// 角色创建
-        /// </summary>
-        /// <param name="worldPos">位置</param>
-        /// <returns>对象</returns>
-        GameObject Create(Vector3 worldPos = default);
-    }
 }
