+++ Scripts/2D/Character/ICharacterState.cs
@@ -1,0 +1,23 @@
+﻿namespace LAB2D
+{
+    /// <summary>
+    /// 角色状态基
+    /// </summary>
+    public interface ICharacterState
+    {
+        /// <summary>
+        /// 角色
+        /// </summary>
+        public void OnEnter();
+
+        /// <summary>
+        /// 当前状态运行
+        /// </summary>
+        public void OnUpdate();
+
+        /// <summary>
+        /// 退出执行
+        /// </summary>
+        public void OnExit();
+    }
+}
